backend/requirements.txt
fastapi 
uvicorn 
python-dotenv
pydantic 
requests
anthropic
python-multipart
python-json-logger
flask
flask-cors